+++ python/bartons/indicators/cci.py
@@ -4,7 +4,7 @@ from polars.plugins import register_plugin_function
 
 from ..prelude import PLUGIN_PATH, wrap_src_indicator
 from ..typing import IntoExprColumn
-from .typprice import TYPPRICE
+from .price import TYPPRICE
 
 
 @wrap_src_indicator